+++ Assets/Part1-HelloWorld/1.5a-TakingATurn/Combat/Attacking.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using Unity.Entities;
+using UnityEngine;
+
+namespace RLTKTutorial.Part1_5A
+{
+	public struct Attacking : IComponentData
+	{
+		public Entity target;
+	} 
+}