+++ Assets/Part1-HelloWorld/1.5a-TakingATurn/Map/MapState/CollidableStateComponent.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Unity.Entities;
+using Unity.Mathematics;
+
+namespace RLTKTutorial.Part1_5A
+{
+	public struct CollidableStateComponent : ISystemStateComponentData
+	{
+		public int2 prevPos;
+	} 
+}