+++ Assets/Part1-HelloWorld/1.5a-TakingATurn/Map/MapState/MapState.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using Unity.Entities;
+using UnityEngine;
+
+namespace RLTKTutorial.Part1_5A
+{
+    public struct MapState : IBufferElementData
+    {
+        public bool blocked;
+        public Entity content;
+    } 
+}