+++ Assets/Part1-HelloWorld/1.5a-TakingATurn/Rendering/RenderableAuthoring.cs
@@ -5,7 +5,6 @@ using UnityEngine;
 
 namespace RLTKTutorial.Part1_5A
 {
-    
     public struct Renderable : IComponentData
     {
         public byte glyph;