+++ Assets/Part1-HelloWorld/1.5a-TakingATurn/TurnSystem/TakingATurn.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using Unity.Entities;
+using UnityEngine;
+
+namespace RLTKTutorial.Part1_5A
+{
+	public struct TakingATurn : IComponentData
+	{ } 
+}