+++ Assets/Part1-HelloWorld/1.5a-TakingATurn/Visibility/VisibilitySystem.cs
@@ -11,7 +11,7 @@ namespace RLTKTutorial.Part1_5A
 {
     [DisableAutoCreation]
     [UpdateAfter(typeof(GenerateMapSystem))]
-    public class VisibilitySystem : JobComponentSystem
+    public class VisibilitySystem : SystemBase
     {
         EntityQuery _mapQuery;
 
@@ -27,85 +27,86 @@ namespace RLTKTutorial.Part1_5A
             RequireForUpdate(_mapQuery);
         }
 
-        protected override JobHandle OnUpdate(JobHandle inputDeps)
+        protected override void OnUpdate()
         {
             var mapEntity = _mapQuery.GetSingletonEntity();
             var map = EntityManager.GetBuffer<MapTiles>(mapEntity);
             var mapData = EntityManager.GetComponentData<MapData>(mapEntity);
 
             if (map.Length == 0 )
-                return inputDeps;
+                return;
 
             bool mapRegenerated = EntityManager.HasComponent<GenerateMap>(mapEntity);
             
-            inputDeps = ProcessViewMemoryEntities(mapRegenerated, map, mapData, inputDeps);
+            ProcessViewMemoryEntities(mapRegenerated, map, mapData);
 
-            inputDeps = ProcessViewEntities(mapRegenerated, map, mapData, inputDeps);
-            
-            return inputDeps;
+            ProcessViewEntities(mapRegenerated, map, mapData);
         }
 
-        JobHandle ProcessViewEntities(bool mapRegenerated, DynamicBuffer<MapTiles> map, MapData mapData, JobHandle inputDeps)
+        void ProcessViewEntities(bool mapRegenerated, DynamicBuffer<MapTiles> map, MapData mapData)
         {
             // Process Entities with only a view and no memory
-            return Entities
-                .WithReadOnly(map)
+            Entities
                 .WithChangeFilter<Position>()
+                .WithReadOnly(map)
                 .WithNone<TilesInMemory>()
                 .ForEach((ref DynamicBuffer<TilesInView> view, in Position pos, in ViewRange range) =>
                 {
                     if (view.Length != map.Length || mapRegenerated)
                         view.ResizeUninitialized(map.Length);
 
+                    var viewArray = view.Reinterpret<bool>().AsNativeArray();
+
                     // Always reset view before rebuilding
                     for (int i = 0; i < view.Length; ++i)
-                        view[i] = false;
+                        viewArray[i] = false;
 
                     var visibility = new VisibilityMap(
                         mapData.width,
                         mapData.height,
                         map.Reinterpret<TileType>().AsNativeArray(),
-                        view.Reinterpret<bool>().AsNativeArray()
+                        viewArray
                         );
 
                     FOV.Compute(pos, range, visibility);
-                }).Schedule(inputDeps);
+                }).ScheduleParallel();
         }
 
-        JobHandle ProcessViewMemoryEntities(bool mapRegenerated, DynamicBuffer<MapTiles> map, MapData mapData, JobHandle inputDeps)
+        void ProcessViewMemoryEntities(bool mapRegenerated, DynamicBuffer<MapTiles> map, MapData mapData)
         {
             // Process Entities with both a view and a memory
-            return Entities
+            Entities
                 .WithChangeFilter<Position>()
                 .WithReadOnly(map)
-                .ForEach((ref DynamicBuffer<TilesInView> view, ref DynamicBuffer<TilesInMemory> memory,
+                .ForEach((
+                ref DynamicBuffer<TilesInView> view, ref DynamicBuffer<TilesInMemory> memory,
                 in Position pos, in ViewRange range) =>
                 {
-
                     if (memory.Length != map.Length || mapRegenerated)
                     {
                         memory.ResizeUninitialized(map.Length);
                         for (int i = 0; i < memory.Length; ++i)
                             memory[i] = false;
                     }
-                    
 
                     if (view.Length != map.Length || mapRegenerated)
                         view.ResizeUninitialized(map.Length);
 
+                    var viewArray = view.Reinterpret<bool>().AsNativeArray();
+
                     // Always reset view before rebuilding
                     for (int i = 0; i < view.Length; ++i)
-                        view[i] = false;
+                        viewArray[i] = false;
 
                     var visibility = new VisibilityMap(
                         mapData.width, mapData.height,
                         map.Reinterpret<TileType>().AsNativeArray(),
-                        view.Reinterpret<bool>().AsNativeArray(),
+                        viewArray,
                         memory.Reinterpret<bool>().AsNativeArray()
                         );
 
                     FOV.Compute(pos, range, visibility);
-                }).Schedule(inputDeps);
+                }).ScheduleParallel();
         }
     }
 }